Restructure contrib directory
