+++ tests/Unit/Contrib/OtlpGrpc/OTLPGrpcExporterTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace OpenTelemetry\Tests\Unit\Contrib;
+namespace OpenTelemetry\Tests\Unit\Contrib\OtlpGrpc;
 
 use AssertWell\PHPUnitGlobalState\EnvironmentVariables;
 use Grpc\UnaryCall;
