+++ tests/Unit/Contrib/Zipkin/ZipkinExporterTest.php
@@ -2,17 +2,16 @@
 
 declare(strict_types=1);
 
-namespace OpenTelemetry\Tests\Unit\Contrib;
+namespace OpenTelemetry\Tests\Unit\Contrib\Zipkin;
 
-use OpenTelemetry\Contrib\ZipkinToNewrelic\Exporter;
+use OpenTelemetry\Contrib\Zipkin\Exporter;
 
 /**
- * @covers OpenTelemetry\Contrib\ZipkinToNewrelic\Exporter
+ * @covers OpenTelemetry\Contrib\Zipkin\Exporter
  */
-class ZipkinToNewrelicExporterTest extends AbstractHttpExporterTest
+class ZipkinExporterTest extends AbstractHttpExporterTest
 {
-    protected const EXPORTER_NAME = 'test.zipkinToNR';
-    protected const LICENSE_KEY = 'abc123';
+    protected const EXPORTER_NAME = 'test.zipkin';
 
     /**
      * @psalm-suppress PossiblyInvalidArgument
@@ -22,7 +21,6 @@ class ZipkinToNewrelicExporterTest extends AbstractHttpExporterTest
         return new Exporter(
             self::EXPORTER_NAME,
             $dsn,
-            self::LICENSE_KEY,
             $this->getClientInterfaceMock(),
             $this->getRequestFactoryInterfaceMock(),
             $this->getStreamFactoryInterfaceMock()
